requirements.txt
openai
streamlit
langchain
faiss-cpu
pypdf
pandas
python-dotenv
tiktoken
PyMuPDF